First Commit
The Arduino_ATMega folder contains all of the files given to us for
this assignment to program the board. This commit also contains
Lab2/program.c which is what we have as far as parts 0 and 1 of Lab2.
This probably doesn’t work because of the source files in the makefile
don’t reference the correct file… yet.